refactor: PostDetail 컴포넌트 위치 이동
