+++ client/src/components/board/index.tsx
@@ -1,2 +1,3 @@
 export { default as PostContent } from './PostContent';
 export { default as Loading } from './Loading';
+export { default as PostDetail } from './PostDetail';
