+++ client/src/components/common/Post/index.tsx
@@ -1,5 +1,4 @@
 export { default as Post } from './post';
-export { default as PostDetail } from './post-detail';
 export { default as Comment } from './comment';
 export { default as CommentUpload } from './comment-upload';
 export { default as DeleteButton } from './delete-button';
