+++ client/src/components/index.tsx
@@ -2,12 +2,7 @@ export { default as Advertisement } from '@/components/common/Advertisement';
 export { default as Layout } from '@/components/base/Layout';
 export { default as MenuView } from '@/components/common/MenuView';
 export { default as Navigator } from '@/components/common/Navigator';
-export {
-  Post,
-  PostDetail,
-  Comment,
-  CommentUpload,
-} from '@/components/common/Post';
+export { Post, Comment, CommentUpload } from '@/components/common/Post';
 export { default as QuillEditBox } from '@/components/common/QuillEditBox';
 export { default as Divider } from '@/components/common/Divider';
 export { default as Avatar } from '@/components/common/Avatar';
